requirements.txt
asgiref==3.7.2
backports.entry-points-selectable==1.1.0
distlib==0.3.3
Django==4.2.4
filelock==3.3.1
pip==21.2.3
platformdirs==2.4.0
setuptools==57.4.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
virtualenv==20.8.1
djangorestframework==3.14.0
